+++ tensorflow/core/kernels/mkl/mkl_layer_norm_op.cc
@@ -15,7 +15,6 @@ limitations under the License.
 
 #ifdef INTEL_MKL
 
-#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 #include "dnnl.hpp"
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow/core/framework/register_types.h"
@@ -23,6 +22,7 @@ limitations under the License.
 #include "tensorflow/core/framework/tensor_types.h"
 #include "tensorflow/core/util/mkl_util.h"
 #include "tensorflow/core/util/tensor_format.h"
+#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 
 using CPUDevice = Eigen::ThreadPoolDevice;
 using dnnl::layer_normalization_forward;
@@ -53,8 +53,8 @@ class MklLayerNormOp : public OpKernel {
       OP_REQUIRES(ctx, shift_tensor.dims() == 1,
                   errors::InvalidArgument("offset must be 1D tensor",
                                           shift_tensor.shape().DebugString()));
-      size_t num_elements_scale = scale_tensor.dim_size(0);
-      size_t num_elements_shift = shift_tensor.dim_size(0);
+      int64_t num_elements_scale = scale_tensor.dim_size(0);
+      int64_t num_elements_shift = shift_tensor.dim_size(0);
       OP_REQUIRES(
           ctx, num_elements_scale == num_elements_shift,
           errors::InvalidArgument("Number of elements in scale and shift",
